Added basic support for an interactive tutorial.
tryclojure.js contains a pages array where tutorial pages can be added. When the user types 'tutorial' at the console it will render page 1 which expects them to enter a specific form.

core.clj was updated with a /eval.json handler. This handler returns a json object with the expression and result. This json object is used by the pages code to determine if the user correctly answered the tutorial page.

The management of history was moved to the middleware function 'wrap-post-history. This way your handler can add anything it wants to the history and this middleware will ensure that only state-modifying forms remain. Current history max is 5 items.